% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{allsolid_incidence_Japan2018_5}
\alias{allsolid_incidence_Japan2018_5}
\title{All solid cancer incidence rates in Japan 2018 (5 year intervals)}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{age}{ages (2.5-97.5)}
  \item{male}{age-specific male rates (per 1 person-year)}
  \item{female}{age-speciic female rates (per 1 person-year)}
  ...
}
}
\source{
\url{https://ganjoho.jp/reg_stat/statistics/data/dl/en.html}
}
\usage{
allsolid_incidence_Japan2018_5
}
\description{
A data set containing the sex- and age-specific all solid cancer incidence
rates in Japan in year 2018 with 5 year age intervals.
}
\keyword{datasets}
