% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{CER}
\alias{CER}
\title{Calculate the cumulated excess risk due to radiation exposure}
\usage{
CER(
  agex,
  doseGy,
  sex,
  riskmodel = LSS_allsolid_mortality,
  wgt = c(1, 0),
  baseline = allsolid_mortality_Japan2018,
  mortality = mortality_Japan2018,
  maxage = max(mortality$age),
  alpha = 0.05,
  n_mcsamp = 10000,
  seed = NULL
)
}
\arguments{
\item{agex}{age(s) at exposure (a single value or a vector)}

\item{doseGy}{dose in Gy or Sv (a single value or a vector)}

\item{sex}{sex 1:male 2:female}

\item{riskmodel}{risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.}

\item{wgt}{weights for ERR vs EAR transfer (e.g. c(1,0) (default), which indicates ERR transfer)}

\item{baseline}{age- and sex-specific baseline rate (data.frame with columns for age, male, female); default: all solid cancer mortality rate of Japan in 2018}

\item{mortality}{age- and sex-specific all cause mortality rate (data.frame with columns for age, male, female); default: all cause mortality rate of Japan in 2018}

\item{maxage}{maximum age to follow up (an integer value)}

\item{alpha}{significance level for the confidence interval (default value = 0.05, which corresponds to 95\% confidence interval)}

\item{n_mcsamp}{number of Monte Carlo sample size (default:10000)}

\item{seed}{random number seed (a single value; default is no seed)}
}
\value{
risk information(vector)
}
\description{
Calculate the cumulated excess risk due to radiation exposure
}
\examples{
   # The following examples use default data provided in the JHPSrisk package
   # for riskmodel (LSS R14 all solid cancer model),
   #     baseline (all solid cancer mortality rates in Japan 2018) and
   #     mortality  (all cause mortality rates in Japan 2018)

   # Cumulated excess risk for male exposed to 0.1 Gy at age 10
   #  followed up to age 90 with err transfer
   CER( agex=10, doseGy=0.1, sex=1, maxage=90 )

   # Cumulated excess risk for female exposed to 0.01 Gy/year at ages 10-19
   #  followed up to age 100 with 7:3 weights for ERR and EAR transfers
   CER( agex=10:19, doseGy=rep(0.01,10), sex=2, maxage=100, wgt=c(.7,.3) )

}
