% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskModel_LSSAllSolidCancerIncidence.r
\docType{data}
\name{LSS_allsolid_incidence}
\alias{LSS_allsolid_incidence}
\title{All solid cancer incidence risk models from Life Span Study}
\format{
A list object with 2 sub list objects for risk model information
         for excess relative risk (err) and excess absolute risk (ear) models.
\describe{
  \item{$err}{
      \describe{
              \item{$para}{a vector of parameter values}
              \item{$var}{ a variance covariance matrix}
              \item{$f}{a function to compute the excess relative risk}
       }
     }
  \item{$ear}{
      \describe{
              \item{$para}{a vector of parameter values}
              \item{$var}{ a variance covariance matrix}
              \item{$f}{a function to compute the excess absolute risk}
              }
       }
}
}
\usage{
LSS_allsolid_incidence
}
\description{
All solid cancer incidence risk models derived from LSS 1958-2009
 (Grant et al., Radiat Res, 2017) for use in LAR function
}
\keyword{datasets}
