% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskModel_LSSAllSolidCancerMortality.r
\docType{data}
\name{LSS_allsolid_mortality}
\alias{LSS_allsolid_mortality}
\title{All solid cancer risk mortality models from Life Span Study Report 14.}
\format{
A list object with 2 sub list objects for risk model information
         for excess relative risk (err) and excess absolute risk (ear) models.
\describe{
  \item{$err}{
      \describe{
              \item{$para}{a vector of parameter values}
              \item{$var}{ a variance covariance matrix}
              \item{$f}{a function to compute the excess relative risk}
       }
     }
  \item{$ear}{
      \describe{
              \item{$para}{a vector of parameter values}
              \item{$var}{ a variance covariance matrix}
              \item{$f}{a function to compute the excess absolute risk}
              }
       }
}
}
\usage{
LSS_allsolid_mortality
}
\description{
All solid cancer risk mortality models derived from LSS Report 14 1950-2003
 (Ozasa et al., Radiat Res, 2012) for use in LAR function
}
\keyword{datasets}
