% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mortality_Japan2018}
\alias{mortality_Japan2018}
\title{All cause mortality rates in Japan 2018}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{age}{ages (1-100)}
  \item{male}{age-specific male rates (per 1 person-year)}
  \item{female}{age-speciic female rates (per 1 person-year)}
  ...
}
}
\source{
\url{https://www.e-stat.go.jp/en}
}
\usage{
mortality_Japan2018
}
\description{
A data set containing the sex- and age-specific all cause mortality rates
in Japan in year 2018.
}
\keyword{datasets}
